Create Data_Generation
There is three dataframe creation process

First one is for restaurants, it contains id of the restaurant, cousine of the restaurant, and flavors of the restaurant,ambiance and profile of the customers that visits . The  flavors are  conditioned on the cuisine. What I mean is, if the cuisine is Italian the flavors contain pizza, pasta in a way that  Italian's are good at. 

Second one is for customer dataset, it contains id of the customer, interest of the customer, ambience that the customer likes , cuisinese and flavor preferences of the customer. 

Third one, is  the reviews of the restaurants in first dataframe which are visited by the customers in the second data frame. Based on the preferences in customer dataframe , the ratings are biased and the habit of going to a restaurant is biased. Customers rate ambiance , taste , service, satisfaction up to 10 points. The similar customers are inclined to go to similar restaurants so the interest and profile section assists on the review section in this part.